Fix misc MacVim warnings and treat warnings as errors in CI
In CI, turn on warnings as errors, but ignore deprecated warnings as we
are still using NSConnection right now and we won't be fixing that for
now. Don't turn on warnings as errors for development (similar to Vim
itself) / outside of CI, because it could make it annoying to build
MacVim locally and across different Xcode versions.

Fix the misc warnings. A lot of 64/32-bit warnings due to careless casts
of NSInteger/NSUInteger.

Also fix up MacVimTests so the waiting for Vim window is more robust
when waiting for 2 windows in a row in vim tutor. Otherwise sometimes
the tests would randomly fail in CI.

+++ src/MacVim/MMApplication.m
@@ -21,7 +21,7 @@
 - (void)sendEvent:(NSEvent *)event
 {
     NSEventType type = [event type];
-    unsigned flags = [event modifierFlags];
+    NSUInteger flags = [event modifierFlags];
 
     // HACK! Intercept 'help' key presses and clear the 'help key flag', else
     // Cocoa turns the mouse cursor into a question mark and goes into 'context
